make epochs cli arg

+++ image-detection/RAID.sh
@@ -3,8 +3,7 @@
 echo "Welcome to RAID: Reese's Awesome Image Detection"
 
 # clean up these stupid export files
-if [ -fe roboflow/README.dataset.txt ]
-then 
+if [ -f "roboflow/README.dataset.txt" ]; then 
     rm roboflow/README.dataset.txt
     rm roboflow/README.roboflow.txt
 fi 
@@ -12,8 +11,8 @@ fi
 
 usage () {
     echo "Usage:"
-    echo " --train             Re-train the death star detector"
-    echo " --dir [path]        Run the image classifier on every file in a directory"
+    echo " --train [num_epochs]     Re-train the death star detector with number of epochs"
+    echo " --dir [path]             Run the image classifier on every file in a directory"
 }
 
 if [ $# -eq 0 ]; then
@@ -25,16 +24,23 @@ fi
 
 if [[ "$1" == "--train" ]]; then
 
+    if [ -z "$2" ]; then
+        echo "Error: missing number epochs"
+        usage
+        exit 1
+    fi
+
+    # rm old runs to make sure file path matches on check
     rm -r runs
     echo "Training!"
 
-    python3 src/train-detection.py 5
+    python3 src/train-detection.py $2 # turn this into command line arg
     exit 0
 fi
 
 if [[ "$1" == "--dir" ]]; then
     if [ -z "$2" ]; then # len == 0
-        echo "Ooops no path. Need path."
+        echo "Error: Ooops no path. Need path."
         usage
         # exit 1
     fi
@@ -42,7 +48,6 @@ if [[ "$1" == "--dir" ]]; then
     echo "Classifying images in directory: $2"
 
     python3 src/detection.py
-
 
     exit 0
 fi
